ex2 is complete.
i created a pointer that points to the file file specified in the read me and have added read and write ability. and then i forked the process and within the if statement the parent process finished running so and wrote onto the file successfully. and then the child process ran and also wrote onto the file successfully. both strings persisted onto the file.