ex3 completed.
initialized two character pointers that point to strings 'Hello' and 'Goodbye'. and then forked the process to have the child print hello and the parent goodbye but also implemented the waitpid function on the else statement to have the parent process wait on the process id specified in the first parameter to finish running.