ex4 completed.
added the exit function to some files i forgot to add them in. create a bin directory with a compiled ls file and just prints a string to the terminal. and in the main process i fork the parent process and have the child run first and the child then runs the specified file passed in by argv