stretch is complete.
we managed to created a specified amount of sub processes by adding a second parameter when running the main file. and in every sub process it will read the balance file, get a random amount to withdraw, if the balance is greater than the amount to be withdrawn then it will lock the file, withdraw the amount from balance, unlock the file when that process is complete then close the file.